update: signal handler の位置を調整
