+++ src/tool_signal.h
@@ -1,0 +1,8 @@
+#ifndef TOOL_SIGNAL_H
+#define TOOL_SIGNAL_H
+
+#include "ft_ping.h"
+
+void tool_setup_signals(t_ping_session *session);
+
+#endif /* TOOL_SIGNAL_H */
